backend/requirements.txt
agora_token_builder==1.0.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
defusedxml==0.7.1
Django==5.2.6
django-allauth==65.11.1
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-crispy-forms==2.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
django-js-asset==3.1.2
djangorestframework-simplejwt==5.4.0
bleach==6.2.0
better-profanity==0.7.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
oauthlib==3.3.1
pillow==11.3.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dotenv==1.1.1
python3-openid==3.2.0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
zope.interface==7.2
